fastapi_service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
python-dotenv==1.0.1
pydantic==1.10.12
alembic==1.11.1